feat(db): implement user and product database operations using Knex refactor(db): remove legacy db.js file and split database logic into db-auth.js and db-products.js for better organization refactor(routes): update routes to use new database modules for user and product operations
