+++ routes/auth.js
@@ -11,7 +11,7 @@ const HttpError = require("../utils/HttpError");
 // Joi
 const Joi = require("joi");
 // DB (Knex)
-const { insertUserTransaction } = require("../db/db");
+const { insertUserTransaction } = require("../db/db-auth");
 
 const authRouter = new Router();
