+++ routes/orders.js
@@ -1,7 +1,5 @@
 // Express promise router
 const Router = require("express-promise-router");
-// Node-postgres (pg)
-const db = require("../db/index");
 
 const ordersRouter = new Router();
